+++ ui-preview/src/main/java/l/files/ui/preview/RectCache.java
@@ -9,7 +9,7 @@ import l.files.fs.File;
 final class RectCache extends PersistenceCache<Rect> {
 
     RectCache(File cacheDir) {
-        super(cacheDir, 1);
+        super(cacheDir, (byte) 1);
     }
 
     @Override
